refactor(core): extract common logic to base classes for controllers and hubs
- Introduced `WartBaseController`, consolidating shared functionality from `WartController` and `WartControllerJwt`.
- Simplified both controllers by inheriting from `WartBaseController` to reduce duplication.
- Extracted shared logic from `WartHub` and `WartHubJwt` into a new `WartBaseHub`.
- Both hubs now inherit from `WartBaseHub`, centralizing SignalR logic.
- Updated method signatures and reused existing utilities to enhance maintainability.

This refactoring reduces duplication, enhances modularity, and simplifies future maintenance.
